Add repo interfacefor creating fakes
