+++ app/src/test/java/co/rahulchowdhury/elly/ui/profile/ElephantProfileViewModelTest.kt
@@ -1,0 +1,10 @@
+package co.rahulchowdhury.elly.ui.profile
+
+import org.junit.Before
+
+class ElephantProfileViewModelTest {
+
+    @Before
+    fun setUp() {
+    }
+}
